Re-organize frame buffer class
