+++ GoofyLand/source/Layers/GoofyLandLayer.h
@@ -6,7 +6,6 @@
 #include "Platform/OpenGL/OpenGLShader.h"
 #include "Basement/Renderer/Texture.h"
 #include "Basement/Renderer/VertexArray.h"
-#include "Basement/Renderer/FrameBuffer.h"
 #include "Basement/Renderer/Model.h"
 
 #include <glm/glm.hpp>
